Fix check_offset overflow on 64-bit systems
On 64-bit systems, the result of subtracting two pointers exceeds the
size of int, so casting to int before checking offsets allows defeating
the check using a large positive offset (greater than or equal to
2^32-size).  Using ptrdiff_t instead of the cast avoids this problem.

Fixes: https://bugs.debian.org/850017

+++ wrestool/fileread.c
@@ -35,7 +35,7 @@
 bool
 check_offset(char *memory, int total_size, char *name, void *offset, int size)
 {
-	int need_size = (int) ((char *) offset - memory + size);
+	ptrdiff_t need_size = (char *) offset - memory + size;
 
 	/*debug("check_offset: size=%x vs %x offset=%x size=%x\n",
 		need_size, total_size, (char *) offset - memory, size);*/
